Restructure and add hardware interface directory
