+++ equipment_handler/inc/adcs_io.h
