+++ equipment_handler/inc/adcs_types.h
