+++ equipment_handler/inc/uart_i2c.h
