+++ equipment_handler/src/adcs_handler.c
