+++ equipment_handler/src/adcs_io.c
